requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
cachetools==4.2.4
    # via google-auth
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.9
    # via requests
click==8.0.3
    # via flask
flask==2.0.2
    # via -r requirements.in
google-api-core[grpc]==2.3.2
    # via
    #   google-api-python-client
    #   google-cloud-secret-manager
google-api-python-client==2.33.0
    # via -r requirements.in
google-auth==2.3.3
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.1.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==0.4.6
    # via -r requirements.in
google-cloud-secret-manager==2.8.0
    # via -r requirements.in
googleapis-common-protos[grpc]==1.54.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.12.3
    # via google-cloud-secret-manager
grpcio==1.43.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.43.0
    # via google-api-core
gunicorn==20.1.0
    # via -r requirements.in
httplib2==0.20.2
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.3
    # via requests
itsdangerous==2.0.1
    # via flask
jinja2==3.0.3
    # via flask
libcst==0.3.23
    # via google-cloud-secret-manager
logzero==1.7.0
    # via -r requirements.in
markupsafe==2.0.1
    # via jinja2
mypy-extensions==0.4.3
    # via typing-inspect
oauthlib==3.1.1
    # via requests-oauthlib
proto-plus==1.19.8
    # via google-cloud-secret-manager
protobuf==3.19.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyparsing==3.0.6
    # via httplib2
python-dateutil==2.8.2
    # via -r requirements.in
pyyaml==6.0
    # via libcst
requests==2.26.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==1.3.0
    # via google-auth-oauthlib
rsa==4.8
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
    #   grpcio
    #   python-dateutil
typing-extensions==4.0.1
    # via
    #   libcst
    #   typing-inspect
typing-inspect==0.7.1
    # via libcst
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.7
    # via requests
werkzeug==2.0.2
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
